Switched to a single codebase for both v4 and v6 code.
