+++ server/src/bpf/response.c
@@ -23,47 +23,43 @@ Augsburg-Traceroute. If not, see <https://www.gnu.org/licenses/>.
 #include "resize.h"
 #include "session.h"
 #include "swap_addr.h"
-#include <linux/icmpv6.h>
+#include "ip_generic.h"
+#include "pseudohdr.h"
+#include <linux/icmp.h>
 #include <linux/if_ether.h>
 #include <linux/in.h>
-#include <linux/ipv6.h>
+#include <linux/ip.h>
 #include <linux/types.h>
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_endian.h>
 
-
-static __be32 pseudo_header(struct ipv6hdr *ip, __u16 probe_len, __u8 protocol)
-{
-    __be32 pseudo_hdr = bpf_htons(probe_len);
-
-    for (int i = 0; i < 8; i++) {
-        pseudo_hdr += ip->daddr.in6_u.u6_addr16[i];
-        pseudo_hdr += ip->saddr.in6_u.u6_addr16[i];
-    }
-    pseudo_hdr += bpf_htons(protocol);
-
-    return pseudo_hdr;
-}
-static void response_init_eth_ip(struct ethhdr *eth, struct ipv6hdr *ip,
-                                 struct in6_addr from, struct in6_addr to)
+static void response_init_eth_ip(struct ethhdr *eth, iphdr_t *ip,
+                                 ipaddr_t from, ipaddr_t to)
 {
     swap_addr_ethhdr(eth);
 
     ip->saddr = from;
     ip->daddr = to;
 
-    ip->nexthdr = IPPROTO_ICMPV6;
+#if defined(TRACEROUTE_V4)
+    ip->protocol = G_PROTO_ICMP;
+    ip->ttl = 64;
+    ip->check = 0;
+    ip->check = csum(ip, sizeof(*ip), 0);
+#elif defined(TRACEROUTE_V6)
+    ip->nexthdr = G_PROTO_ICMP;
     ip->hop_limit = 64;
+#endif
 }
 
 static void response_init_icmp(struct session_key *session,
-                               struct icmp6hdr *icmp, union trhdr *tr,
-                               struct trhdr_payload *payload ,probe_error error)
+                               struct icmphdr *icmp, union trhdr *tr,
+                               struct trhdr_payload *payload, probe_error error)
 {
-    icmp->icmp6_type = ICMPV6_ECHO_REPLY;
-    icmp->icmp6_code = 1;
-    icmp->icmp6_dataun.u_echo.identifier = session->identifier;
-    icmp->icmp6_dataun.u_echo.sequence = 0;
+    icmp->type = 0;
+    icmp->code = 1;
+    icmp->un.echo.id = session->identifier;
+    icmp->un.echo.sequence = 0;
 
     tr->response.state = error;
     tr->response.err_msg_len = 0;
@@ -77,13 +73,13 @@ static void response_init_icmp(struct session_key *session,
 
 INTERNAL int response_create_err(struct cursor *cursor,
                                  struct session_key *session, probe_error error,
-                                 struct ethhdr **eth, struct ipv6hdr **ip)
+                                 struct ethhdr **eth, iphdr_t **ip)
 {
-    struct icmp6hdr *icmp;
+    struct icmphdr *icmp;
     union trhdr *tr;
 
-    struct in6_addr dest_addr = session->addr;
-    struct in6_addr source_addr = (**ip).daddr;
+    ipaddr_t dest_addr = session->addr;
+    ipaddr_t source_addr = (**ip).daddr;
 
     __u16 payload_len = sizeof(*icmp) + sizeof(*tr);
 
@@ -98,24 +94,29 @@ INTERNAL int response_create_err(struct cursor *cursor,
     response_init_eth_ip(*eth, *ip, source_addr, dest_addr);
     response_init_icmp(session, icmp, tr, NULL, error);
 
-    icmp->icmp6_cksum = 0;
-    icmp->icmp6_cksum = csum(icmp, payload_len, pseudo_header(*ip, payload_len, IPPROTO_ICMPV6));
+#if defined(TRACEROUTE_V4)
+    __be32 seed = 0;
+#elif defined(TRACEROUTE_V6)
+    __be32 seed = pseudo_header(*ip, payload_len, G_PROTO_ICMP);
+#endif
+    icmp->checksum = 0;
+    icmp->checksum = csum(icmp, payload_len, seed);
 
     return 0;
 }
 
 INTERNAL int response_create(struct cursor *cursor, struct session_key *session,
                              struct session_state *state, struct ethhdr **eth,
-                             struct ipv6hdr **ip)
+                             iphdr_t **ip)
 {
-    struct icmp6hdr *icmp;
+    struct icmphdr *icmp;
     union trhdr *tr;
     struct trhdr_payload *payload;
     __u64 timespan_ns;
 
-    struct in6_addr dest_addr = session->addr;
-    struct in6_addr source_addr = (**ip).daddr;
-    struct in6_addr from_addr = (**ip).saddr;
+    ipaddr_t dest_addr = session->addr;
+    ipaddr_t source_addr = (**ip).daddr;
+    ipaddr_t from_addr = (**ip).saddr;
 
     __u16 payload_len = sizeof(*icmp) + sizeof(*tr) + sizeof(*payload);
 
@@ -130,7 +131,15 @@ INTERNAL int response_create(struct cursor *cursor, struct session_key *session,
     if (PARSE(cursor, &payload) < 0)
         return -1;
 
+#if defined(TRACEROUTE_V4)
+    for (int i = 0; i < 5; i++)
+        payload->addr.in6_u.u6_addr16[i] = 0;
+
+    payload->addr.in6_u.u6_addr16[5] = 0xffff;
+    payload->addr.in6_u.u6_addr32[3] = from_addr;
+#elif defined(TRACEROUTE_V6)
     payload->addr = from_addr;
+#endif
 
     // Calculate timestamp.
     timespan_ns = bpf_ktime_get_ns() - state->timestamp_ns;
@@ -139,7 +148,13 @@ INTERNAL int response_create(struct cursor *cursor, struct session_key *session,
     response_init_eth_ip(*eth, *ip, source_addr, dest_addr);
     response_init_icmp(session, icmp, tr, payload, 0);
 
-    icmp->icmp6_cksum = 0;
-    icmp->icmp6_cksum = csum(icmp, payload_len, pseudo_header(*ip, payload_len, IPPROTO_ICMPV6));
+#if defined(TRACEROUTE_V4)
+    __be32 seed = 0;
+#elif defined(TRACEROUTE_V6)
+    __be32 seed = pseudo_header(*ip, payload_len, G_PROTO_ICMP);
+#endif
+    icmp->checksum = 0;
+    icmp->checksum = csum(icmp, payload_len, seed);
+
     return 0;
 }
