+++ server/src/bpf/session.h
@@ -17,12 +17,27 @@ You should have received a copy of the GNU General Public License along with
 Augsburg-Traceroute. If not, see <https://www.gnu.org/licenses/>.
 */
 
-#ifndef INTERNAL_H
-#define INTERNAL_H
+#ifndef SESSION_H
+#define SESSION_H
 
+#include "internal.h"
+#include "ip_generic.h"
+#include <linux/bpf.h>
 #include <linux/types.h>
-#include <bpf/bpf_helpers.h>
 
-#define INTERNAL __hidden
+struct session_key {
+    ipaddr_t addr;
+    __u16 identifier;
+    __u16 padding;
+};
+
+struct session_state {
+    __u64 timestamp_ns;
+};
+
+INTERNAL int session_delete(struct session_key *session);
+INTERNAL struct session_state *session_find(struct session_key *key);
+INTERNAL int session_add(struct session_key *session,
+                         struct session_state *state);
 
 #endif
